# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/insumos": {
            "get": {
                "tags": [
                    "Insumos"
                ],
                "summary": "Obtener todos los insumos",
                "operationId": "e7e959436667829a4acb2b85f7999905",
                "responses": {
                    "200": {
                        "description": "Listado de insumos encontrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseInsumoList"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron tratamientos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontraron tratamientos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Insumos"
                ],
                "summary": "Crear un nuevo insumo",
                "operationId": "ca273dd445b075888b9c1c71eb9d28fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "componentes",
                                    "precio_insumo",
                                    "cantidad",
                                    "nombre",
                                    "fecha_expiracion"
                                ],
                                "properties": {
                                    "componentes": {
                                        "type": "string",
                                        "example": "Ácido hialurónico"
                                    },
                                    "precio_insumo": {
                                        "type": "number",
                                        "example": 4500.5
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Insumo facial A"
                                    },
                                    "fecha_expiracion": {
                                        "type": "string",
                                        "example": "31/12/2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Insumo creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseInsumoSingle"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de datos"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/insumos/{id}": {
            "get": {
                "tags": [
                    "Insumos"
                ],
                "summary": "Obtener un insumo por ID",
                "operationId": "391b872fc3768f3953a901eb407ab505",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del insumo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insumo encontrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseInsumoSingle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Insumo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Insumo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Insumos"
                ],
                "summary": "Actualizar un insumo",
                "operationId": "d38de4e319dc5339e3ddd56f730b7bb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del insumo a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "componentes": {
                                        "type": "string",
                                        "example": "Nuevo componente"
                                    },
                                    "precio_insumo": {
                                        "type": "number",
                                        "example": 3000
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Insumo actualizado"
                                    },
                                    "fecha_expiracion": {
                                        "type": "string",
                                        "example": "01/01/2026"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Insumo actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseInsumoSingle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Insumo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Insumo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de datos"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Insumos"
                ],
                "summary": "Eliminar un insumo",
                "operationId": "6e543997fc4abd121d1461a82b9be341",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del insumo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insumo eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Insumo eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Insumo no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Insumo no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Insumo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "componentes": {
                        "type": "string",
                        "example": "Ácido hialurónico"
                    },
                    "precio_insumo": {
                        "type": "number",
                        "example": 4500.5
                    },
                    "cantidad": {
                        "type": "integer",
                        "example": 10
                    },
                    "cantidad_min": {
                        "type": "integer",
                        "example": 5
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Insumo facial A"
                    },
                    "fecha_expiracion": {
                        "type": "string",
                        "example": "2025-12-31"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-04T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-04T12:00:00Z"
                    },
                    "alertaBajoStock": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "ResponseInsumoList": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Insumos encontrados correctamente"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Insumo"
                        }
                    }
                },
                "type": "object"
            },
            "ResponseInsumoSingle": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Insumo encontrado correctamente"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Insumo"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}